+++ simpleplot/ploting/plot_data_types/step_data.py
@@ -55,7 +55,7 @@ class StepData(PlotData, SessionNode):
         for i,value in enumerate(self._axes):
             if value in kwargs.keys():
                 if isinstance(kwargs[value],np.ndarray) or isinstance(kwargs[value],list):
-                    elements[i] = np.array(kwargs[value])
+                    elements[i] = np.array(kwargs[value]).astype(np.float)
                     changed[self._axes.index(value)] = True
 
         if elements[self._axes.index('z')] is None:
